requirements.txt
Flask==1.1.2
selenium==3.141.0

PyYAML~=6.0